Set up all initial fires
